+++ previous-versions/common/src/codatcommon/models/operations/get_profile_syncsettings.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import syncsettings as shared_syncsettings
+from ...models.shared import syncsettings as shared_syncsettings
 from typing import Optional
 
 
@@ -12,12 +11,10 @@ from typing import Optional
 class GetProfileSyncSettingsResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""
     sync_settings: Optional[shared_syncsettings.SyncSettings] = dataclasses.field(default=None)
     r"""OK"""