+++ previous-versions/common/src/codatcommon/models/shared/pushoption.py
@@ -2,10 +2,10 @@
 
 from __future__ import annotations
 import dataclasses
-from ..shared import pushoptionchoice as shared_pushoptionchoice
-from ..shared import pushoptionproperty as shared_pushoptionproperty
-from ..shared import pushoptiontype as shared_pushoptiontype
-from ..shared import pushvalidationinfo as shared_pushvalidationinfo
+from .pushoptionchoice import PushOptionChoice
+from .pushoptionproperty import PushOptionProperty
+from .pushoptiontype import PushOptionType
+from .pushvalidationinfo import PushValidationInfo
 from codatcommon import utils
 from dataclasses_json import Undefined, dataclass_json
 from typing import Dict, List, Optional
@@ -18,12 +18,12 @@ class PushOption:
     r"""The property's display name."""
     required: bool = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('required') }})
     r"""The property is required if `True`."""
-    type: shared_pushoptiontype.PushOptionType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    type: PushOptionType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
     r"""The option type."""
     description: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('description'), 'exclude': lambda f: f is None }})
     r"""A description of the property."""
-    options: Optional[List[shared_pushoptionchoice.PushOptionChoice]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('options') }})
-    properties: Optional[Dict[str, shared_pushoptionproperty.PushOptionProperty]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('properties') }})
-    validation: Optional[shared_pushvalidationinfo.PushValidationInfo] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('validation'), 'exclude': lambda f: f is None }})
+    options: Optional[List[PushOptionChoice]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('options') }})
+    properties: Optional[Dict[str, PushOptionProperty]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('properties') }})
+    validation: Optional[PushValidationInfo] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('validation'), 'exclude': lambda f: f is None }})