+++ previous-versions/common/src/codatcommon/models/shared/supportedfeature.py
@@ -2,8 +2,8 @@
 
 from __future__ import annotations
 import dataclasses
-from ..shared import featurestate as shared_featurestate
-from ..shared import featuretype as shared_featuretype
+from .featurestate import FeatureState
+from .featuretype import FeatureType
 from codatcommon import utils
 from dataclasses_json import Undefined, dataclass_json
 
@@ -11,9 +11,9 @@ from dataclasses_json import Undefined, dataclass_json
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class SupportedFeature:
-    feature_state: shared_featurestate.FeatureState = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('featureState') }})
+    feature_state: FeatureState = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('featureState') }})
     r"""The current release state of the feature."""
-    feature_type: shared_featuretype.FeatureType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('featureType') }})
+    feature_type: FeatureType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('featureType') }})
     r"""The type of feature."""